moved the tests into their respective test files. Fixed castArray tests
